Cmpctmalloc: Use size of previous memory area.
This is a change to the internal implementation of cmpctmalloc
to store the size of the previous area in memory order instead of a
pointer to the start of the previous area.  This is a precursor
to reducing the allocation header size fields to 2 bytes, reducing
the header from 8 bytes to 4 bytes, since allocations are always a lot
less than 64k.

Also adds ifdefs so that cmpctmalloc can be compiled standalone and
tests can be run.