Waste less memory in malloc in the 4-16k range.
Previously we rounded up to nearest 4k page for allocations above about
3.9k.  Now we have an intermediate area from 4-16k where we create a
larger malloc arena sized 8, 12, or 16k, and then do mallocs from there.
The hope is that we waste less, but the risk is that a single small
allocation after the big one keeps the 16k from being returned to the
page allocator.

Alternatively we could consider not using this new scheme for
allocations between 7-8k, 11-12k, and 15-16k, where we would only waste
1k with the old page-aligned strategy.

We continue to use the page allocator for medium sizes that are exact
multiples of 4k, ie 4k, 8k and 12k exactly.