Aligned allocation semantics for esp-idf 4.3
Starting in esp-idf 4.3 you can call regular free()
for memory allocated using the aligned allocation
API.  This means we can't use the old trick of
storing the location of the real allocation in a
word immediately preceeding the aligned address.
Instead we have to slice the free-list entry in 3
slices with free areas to the left and the right of
the aligned allocation.