cmpctmalloc: Mark pages used for smaller allocations differently.
Previously we used the address of the heap structure itself
for the tag when we allocated pages that were to be used
as a malloc arena for smaller allocations.  However if we
want to segregate allocations by tag we would like to move
the tag to the per-page structure, instead of having a tag
on every single allocation.  So we introduce a new page status
instead of using the tag to distinguish between pages that
are allocated directly vs. pages that contain smaller
allocations.